fix(gateway): use INTEGER for telegram_message_id to match entity Int? mapping

+++ gateway/src/main/resources/db/migration/V9__add_telegram_message_id.sql
@@ -1,1 +1,1 @@
-ALTER TABLE telegram_deliveries ADD COLUMN telegram_message_id BIGINT;
+ALTER TABLE telegram_deliveries ADD COLUMN telegram_message_id INTEGER;
